finished Assignment 2

+++ Matchismo/Matchismo/Model/Card.h
@@ -15,6 +15,6 @@
 @property (nonatomic, getter = isChosen) BOOL chosen;
 @property (nonatomic, getter = isMatched) BOOL matched;
 
-- (int)match:(NSArray *)otherCards;
+- (NSInteger)match:(NSArray *)otherCards;
 
 @end
