+++ Matchismo/Matchismo/Model/Card.m
@@ -14,12 +14,11 @@
 
 @implementation Card
 
-- (int)match:(NSArray *)otherCards
+- (NSInteger)match:(NSArray *)otherCards
 {
     int score = 0;
     
-    for (Card* card in otherCards)
-    {
+    for (Card* card in otherCards) {
         if ([card.contents isEqualToString:self.contents])
             score += 1;
     }
